refactor(types): 🚚 move Project model to common/types/models
- Relocated Project, ProjectStatus, User, and RequiredRoles types.
- Centralized type definitions for broader application access.
- Enhanced project structure for better type management.
