+++ src/common/types/index.ts
@@ -1,1 +1,2 @@
 export * from './general';
+export * from './models';
